backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
sentence-transformers
scikit-learn
torch
python-multipart
jinja2
motor
dnspython
python-dotenv
passlib[bcrypt]
python-jose[cryptography]

